requirements.txt
asgiref==3.8.1
config==0.5.1
django-allauth==65.4.1
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.10
inflection==0.5.1
joblib==1.5.0
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
